update Native UInt

+++ kfswatch/src/mingwX64Main/kotlin/io/github/irgaly/kfswatch/internal/platform/FileWatcher.kt
@@ -334,10 +334,10 @@ internal actual class FileWatcher actual constructor(
                                     val handle = checkNotNull(
                                         CreateFileW(
                                             lpFileName = targetPath.originalPath,
-                                            dwDesiredAccess = FILE_LIST_DIRECTORY,
+                                            dwDesiredAccess = FILE_LIST_DIRECTORY.toUInt(),
                                             dwShareMode = (FILE_SHARE_READ or FILE_SHARE_WRITE or FILE_SHARE_DELETE).toUInt(),
                                             lpSecurityAttributes = null,
-                                            dwCreationDisposition = OPEN_EXISTING,
+                                            dwCreationDisposition = OPEN_EXISTING.toUInt(),
                                             dwFlagsAndAttributes = (FILE_FLAG_BACKUP_SEMANTICS or FILE_FLAG_OVERLAPPED).toUInt(),
                                             hTemplateFile = null
                                         )
@@ -565,7 +565,7 @@ internal actual class FileWatcher actual constructor(
     actual fun pause() {
         val result = WaitForSingleObject(
             hHandle = threadSemaphoreHandle,
-            dwMilliseconds = 0
+            dwMilliseconds = 0U
         )
         if (result == WAIT_OBJECT_0) {
             // ロックされていない状態からロックされた
@@ -659,9 +659,9 @@ internal actual class FileWatcher actual constructor(
                         ).toUInt(),
                 lpSource = null,
                 dwMessageId = errorCode,
-                dwLanguageId = 0,
+                dwLanguageId = 0U,
                 lpBuffer = messagePointer.ptr.reinterpret(),
-                nSize = 0,
+                nSize = 0U,
                 Arguments = null
             )
             val message = checkNotNull(messagePointer.value).toKString()
